Single pixel drawing

+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod canvas;
 
-pub type Point = (f32, f32);
+pub type Point = (u32, u32);
 pub type Triangle = (Point, Point, Point);
